moving influxd to independant package to assist with intgration tests
